% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_pdf}
\alias{theme_pdf}
\title{Pretty theme}
\usage{
theme_pdf(show.ticks = TRUE, legend.pos = NULL)
}
\arguments{
\item{show.ticks}{Show x- and y- ticks.}

\item{legend.pos}{Vector with x and y position of the legend.}
}
\description{
Pretty theme
}
\examples{
ggplot2::ggplot() + geom_point(aes(x=rnorm(1000), y=rnorm(1000), color=(1:1000 / 1000))) + theme_pdf(F, legend.pos=c(1, 1))

}
